pyproject.toml
[tool.poetry]
name = "product-api"
version = "0.1.0"
description = "REST API для управления продуктами"
authors = ["mrGremlin2019 <vadim-ivanov01@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
